pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nstd"
version = "0.1.0"
description = "Personal task synchronisation daemon and TUI — syncs GitHub Issues, Jira, Asana, and Google Calendar"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    {name = "nate-double-u"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "PyGithub>=2.1.1",
    "gql[httpx]>=3.5.0",
    "httpx>=0.27.0",
    "jira>=3.8.0",
    "asana>=5.0.0",
    "google-api-python-client>=2.130.0",
    "google-auth-oauthlib>=1.2.0",
    "textual>=0.61.0",
    "keyring>=25.0.0",
    "ollama>=0.2.0",
    "click>=8.1.7",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "responses>=0.25.0",
    "respx>=0.21.0",
    "freezegun>=1.4.0",
    "factory-boy>=3.3.0",
]

[project.scripts]
nstd = "nstd.cli:main"

[project.urls]
Homepage = "https://github.com/nate-double-u/nstd"
Repository = "https://github.com/nate-double-u/nstd"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that require real API credentials (deselected by default)",
]
addopts = "-m 'not integration'"

[tool.coverage.run]
source = ["nstd"]

[tool.coverage.report]
fail_under = 90
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]
